Fix and test for #190
Baffling that this got through, but I obviously botched the impl
of __persistent__ and __persistent__=. They were set up as meta-
class methods but using the instance method versions. So they
received a RubyClass object but then the invoker has to cast to
JavaProxy. Fixed by making them static and adjusting impl
accordingly.